pyproject.toml
[tool.poetry]
name = "ai-agents-tool-calling"
version = "0.1.0"
description = ""
authors = ["Samuel Matioli <samuel.matioli@datastax.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
langchain = "^0.2.16"
langchain-community = "^0.2.16"
astrapy = "^1.4.1"
python-dotenv = "^1.0.1"
openai = "^1.44.0"
langchain-openai = "^0.1.23"

[tool.poetry.scripts]
agent = "ai_agents_tool_calling:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
